feat: add Test History page with search and table as components
